refactor: move WASM patches into _runtime/_wasm package
