+++ marimo/_runtime/_wasm/__init__.py
@@ -1,0 +1,1 @@
+# Copyright 2026 Marimo. All rights reserved.
